+++ Reach/src/utopia/reach/component/factory/contextual/BaseContextualFactory.scala
@@ -11,6 +11,8 @@ import utopia.paradigm.color.Color
 trait BaseContextualFactory[+Repr]
 	extends ContextualFactory[StaticBaseContext, Repr] with StaticBaseContextWrapper[StaticBaseContext, Repr]
 {
+	// IMPLEMENTED  ---------------------------------
+	
 	override def base = context
 	
 	override def current: StaticBaseContext = base